Settings panel: react to key and type errors for touch passthrough settings.
With the inclusion of touch support toggle, there might come a time when touch passthrough keys may present issues such as key error and type error. Typically this happens if a configuration profile other than normal profile is in use and 'touch' section does not have passthrough keys. Thus react to these by:
* Touch passthrough: set value to 5 seconds if touch passthrough duration key is not found.
* Manual touch passthrough toggle: make it a boolean (string to bool conversion).
Note that touch passthrough is hereby deprecated and will be removed in a future add-on release.