Add-on settings panel: remove check for touch interaction flag if touch interaction toggle is built into NVDA.
Until NVDA included ability to toggle touch interaction itself, the add-on was responsible for presenting the toggle checkbox as part of its settings panel. As NVDA 2020.3 or later includes such a checkbox itself, there is no need to provide touch interaction toggle checkbox from the add-on. Besides, settings panel will be shown if older NVDA releases are in use.