Do not register config switch handler nor settings change handler if NVDA 2020.3 or later is in use.
In NVDA 2020.2 and earlier, because NVDA didn't have config switch handler for touch handler (let alone touch support toggle flag), the add-on was responsible for keeping an eye on config profile swithces and enabling/disabling touch support. Now that 2020.3 will include touchHandler.setTouchSupport wihch is employed when switching profiles, do not register the add-on's config profile switch handler if NVDA 2020.3 or later is active. This also extends to add-on settings change handler.