Add a touch command to toggle Windows 10 dictation (Windows+H).
A user request: add a touch gesture to toggle dictation on Windows 10 (Windows+H). For this, four finger flick left is assigned to compliment four finger flick right for showing/hiding touch keyoard. Note that this gesture will be active on Version 1709 and later (otherwise an error message iwll be presented).